feat(eslr): 调整语义动作的挂载逻辑，部分细节优化
调整语义动作的挂载逻辑；
调整部分类型定义；
部分细节优化。

+++ src/common/fa.h
@@ -16,6 +16,6 @@
 
 using namespace std;
 
-typedef int state_id_t;
+typedef size_t state_id_t;
 typedef pair<state_id_t, state_id_t> sub_fa_t;
 typedef unordered_map<char, set<state_id_t>> transition_map_t;
